Create a Linter::Base class and extract common code to it.
